fix: fix error on save customer pets
New test on postman script for validating every request.

Updating customers with pets was causing a 500 response
code and sent a response with this error: "A collection with
cascade=”all-delete-orphan” was no longer referenced
by the owning entity instance".  This was corrected.
The Customer's pets Arraylist was being replace wholesale
and hibernate was losing the reference to the collection.
Found StackOverflow 5587482.  Updating the exsiting list
with each pet kept the reference of the list intact.